Added Create blog feature

+++ back-end/controllers/productController.js
@@ -24,35 +24,46 @@ const handleMultipartData = multer({
 const productController = {
     async store(req, res, next) {
         // Multipart form data
+
         handleMultipartData(req, res, async (err) => {
             if (err) {
                 return next(CustomErrorHandler.serverError(err.message));
             }
-            const filePath = req.file.path;
-            // validation
-            const { error } = productSchema.validate(req.body);
-            if (error) {
-                // Delete the uploaded file
-                fs.unlink(`${appRoot}/${filePath}`, (err) => {
-                    if (err) {
-                        return next(
-                            CustomErrorHandler.serverError(err.message)
-                        );
-                    }
-                });
-
-                return next(error);
-                // rootfolder/uploads/filename.png
+            var filePath;
+            if (req.file) {
+                filePath = req.file.path;
+                // validation
+                const { error } = productSchema.validate(req.body);
+                if (error) {
+                    // Delete the uploaded file
+                    fs.unlink(`${appRoot}/${filePath}`, (err) => {
+                        if (err) {
+                            return next(
+                                CustomErrorHandler.serverError(err.message)
+                            );
+                        }
+                    });
+    
+                    return next(error);
+                    // rootfolder/uploads/filename.png
+                }
             }
-
+           
             const { title, story } = req.body;
             let document;
             try {
-                document = await Product.create({
-                    title,
-                    story,
-                    image: filePath,
-                });
+                if (filePath) {
+                    document = await Product.create({
+                        title,
+                        story,
+                        image: filePath,
+                    });
+                }else{
+                    document = await Product.create({
+                        title,
+                        story,
+                    });
+                }
             } catch (err) {
                 return next(err);
             }
